Refactor audit log functionality: integrate into admin controller, remove separate routes and controller, and update logger actions.
