+++ app-backend/src/controllers/auth.controller.js
@@ -29,7 +29,7 @@ export const register = async (req, res) => {
 
     const newUser = new User({ name, email, password, role, phone, address });
     await newUser.save();
-    await req.audit.log(newUser._id, ACTIONS.PROFILE_UPDATED, { registered: true });
+    await req.audit.log(newUser._id, ACTIONS.PROFILE_CREATED, { registered: true });
     res.status(201).json({ message: 'User registered successfully' });
   } catch (err) {
     res.status(500).json({ message: err.message });
