+++ app-backend/src/middleware/logger.js
@@ -1,7 +1,11 @@
 import AuditLog from '../models/AuditLogs.js';  // your separate model
 
-// Define all your action constants
+const AUDIT_LOG_ENABLED = process.env.AUDIT_LOG_ENABLED === 'true';
+const IS_DEV = process.env.NODE_ENV === 'development';
+
+// Define all action constants
 export const ACTIONS = {
+  PROFILE_CREATED: 'PROFILE_CREATED',
   PROFILE_UPDATED: 'PROFILE_UPDATED',
   LOGIN_SUCCESS: 'LOGIN_SUCCESS',
   SHIFT_APPLIED: 'SHIFT_APPLIED',
@@ -14,11 +18,14 @@ export const ACTIONS = {
 // Middleware to attach audit logging function to req
 export const auditMiddleware = (req, res, next) => {
   req.audit = {
-    log: async (userId, action, details = {}) => {
+    log: async (user, action, details = {}) => {
+      if (IS_DEV) {
+        console.log(`[DEV] Audit log: ${action} by user ${user?._id || user}`, details);
+      }
+      if (!AUDIT_LOG_ENABLED) return;
       try {
-        const logEntry = new AuditLog({ userId, action, details });
+        const logEntry = new AuditLog({ user, action, details });
         await logEntry.save();
-        console.log(`Audit log: ${action} by user ${userId}`, details);
       } catch (err) {
         console.error('Failed to save audit log:', err.message);
       }
