+++ app-backend/src/models/AuditLogs.js
@@ -1,5 +1,6 @@
 import mongoose from 'mongoose';
 
+
 const AuditLogSchema = new mongoose.Schema(
   {
     user: { type: mongoose.Schema.Types.ObjectId, ref: "User", required: false },
@@ -9,6 +10,12 @@ const AuditLogSchema = new mongoose.Schema(
   },
   { versionKey: false }
 );
+
+// Static method to purge logs older than a given number of days
+AuditLogSchema.statics.purgeOldLogs = async function (days) {
+  const cutoff = new Date(Date.now() - days * 24 * 60 * 60 * 1000);
+  return this.deleteMany({ timestamp: { $lt: cutoff } });
+};
 
 AuditLogSchema.index({ user: 1, timestamp: -1 });
 AuditLogSchema.index({ action: 1, timestamp: -1 });
