+++ app-backend/src/routes/index.js
@@ -9,7 +9,6 @@ import userRoutes from './user.routes.js';
 import adminRoutes from './admin.routes.js';
 
 import availabilityRoutes from './availability.routes.js'; 
-import auditLogRoutes from './auditlogs.routes.js';
 
 const router = Router();
 
@@ -21,7 +20,6 @@ router.use('/messages', messageRoutes);
 router.use('/admin', adminRoutes);
 
 router.use('/availability', availabilityRoutes); 
-router.use('/', auditLogRoutes);
 router.use ('/messages', messageRoutes);
 
 router.use('/users', userRoutes);
